pyproject.toml
[project]
name = "rnt-hub"
version = "0.1.0"
description = "Projeto final da EBAC"
authors = [
    {name = "RNT13",email = "renatornt13@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "poetry-plugin-export (>=1.9.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Configurações de ferramentas
[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = "tests.py test_*.py *_tests.py"
pythonpath = "."

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "flake8-isort (>=7.0.0,<8.0.0)",
    "flake8-black (>=0.4.0,<0.5.0)"
]
